requirements.txt
google-generativeai
customtkinter
pyserial
SpeechRecognition
PyAudio
Pillow
requests
pyttsx3
numpy
python-dotenv
